feat: Add new user role 'USER' to Role enum
This commit adds a new value 'USER' to the Role enum in the role.ts file. This allows for better representation of user roles in the application.

Refactor layout and styles

This commit includes several refactorings to the layout and styles of the application. It removes commented out code in BaseLayout.vue, updates z-index values in GlobalSearch and BaseLayout components, and adds hover box shadow to gridTemplateColumns.

Add new columns and default value to Anmeldung table

This commit adds new columns and a default value to the Anmeldung table in the migration.sql files. It adds the 'activatedAt' column to the Anmeldung table and sets a default value of 'AUSSTEHEND' for the 'status' column.

Update T-Shirt Bestellung form

This commit updates the T-Shirt Bestellung form in the FormTShirtBestellungGeneral.vue file. It modifies the styling of the form by changing the margin-bottom value.

Update PersonCreate view

This commit updates the PersonCreate view in the PersonCreate.vue file. It modifies the layout of the view by changing the heading element from a div to an h2 and adds a grid layout to the form.

Update mail sending functionality

This commit updates the mail sending functionality in the sendMail function in the mail.ts file. It modifies the 'to' field to send the mail to a specific email address for testing purposes.

Add new migration for activation token in Anmeldung table

This commit adds a new migration file for the Anmeldung table in the migration.sql files. It adds a new column 'activationToken' to the Anmeldung table and creates a unique index for the column.

Update AnmeldungStatus enum

This commit updates the AnmeldungStatus enum in the AnmeldungStatus.ts file. It adds a new value 'AUSSTEHEND' to represent an Anmeldung that is not yet verified.

Update sendMailConfirmEmailRequest function

This commit updates the sendMailConfirmEmailRequest function in the sendMailConfirmEmailRequest.ts file. It modifies the activationUrl to include the correct URL based on whether the request is for public or private registration.

Update Anmeldung routes

This commit updates the Anmeldung routes in the routes.ts file. It adds new routes for public Anmeldung email verification and confirmation.

Update anmeldungRouter

This commit updates the anmeldungRouter in the anmeldung.router.ts file. It adds a new route for the anmeldungEmailConfirmProcedure.

Update accountEmailConfirmProcedure

This commit adds a new accountEmailConfirmProcedure in the account.ts file. It handles the email confirmation for user accounts.

Update anmeldungEmailConfirmProcedure

This commit adds a new anmeldungEmailConfirmProcedure in the anmeldungEmailConfirm.ts file. It handles the email confirmation for Anmeldung.